evaluation/flow/requirements.txt
# Azure SDK
azure-identity>=1.15.0
azure-search-documents>=11.6.0
openai>=1.12.0

# Prompt Flow
promptflow>=1.8.0
promptflow-tools>=1.4.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# Evaluation
scikit-learn>=1.3.0
